Rename and remove subclass
